Extract common bits of Sysconfig::ConfigWriter to ConfigWriter
