+++ test/y2network/sysconfig/config_writer_test.rb
@@ -95,9 +95,11 @@ describe Y2Network::Sysconfig::ConfigWriter do
 
     let(:routes) { [route, default_route] }
 
-    let(:dns_writer) { instance_double(Y2Network::Sysconfig::DNSWriter, write: nil) }
-    let(:hostname_writer) { instance_double(Y2Network::Sysconfig::HostnameWriter, write: nil) }
-    let(:interfaces_writer) { instance_double(Y2Network::Sysconfig::InterfacesWriter, write: nil) }
+    let(:dns_writer) { instance_double(Y2Network::ConfigWriters::DNSWriter, write: nil) }
+    let(:hostname_writer) { instance_double(Y2Network::ConfigWriters::HostnameWriter, write: nil) }
+    let(:interfaces_writer) do
+      instance_double(Y2Network::ConfigWriters::InterfacesWriter, write: nil)
+    end
     let(:sysctl_config_file) do
       CFA::SysctlConfig.new do |f|
         f.forward_ipv4 = false
@@ -112,12 +114,12 @@ describe Y2Network::Sysconfig::ConfigWriter do
       allow(CFA::RoutesFile).to receive(:new)
         .with(no_args)
         .and_return(routes_file)
-      allow(Y2Network::Sysconfig::DNSWriter).to receive(:new)
+      allow(Y2Network::ConfigWriters::DNSWriter).to receive(:new)
         .and_return(dns_writer)
-      allow(Y2Network::Sysconfig::HostnameWriter).to receive(:new)
+      allow(Y2Network::ConfigWriters::HostnameWriter).to receive(:new)
         .and_return(hostname_writer)
       allow_any_instance_of(Y2Network::Sysconfig::ConnectionConfigWriter).to receive(:write)
-      allow(Y2Network::Sysconfig::InterfacesWriter).to receive(:new)
+      allow(Y2Network::ConfigWriters::InterfacesWriter).to receive(:new)
         .and_return(interfaces_writer)
       allow(CFA::SysctlConfig).to receive(:new).and_return(sysctl_config_file)
       allow(sysctl_config_file).to receive(:load)
